Add new syscall to modify the chosen service backends.
This syscall can be used by clients to modify the selection, the
relative priority and the accompanying parameter structs for the
backends the client wants to use. As it doesn't fit any specific
category of existing Client trait, add it into the basic
PollClient - basically it's a meta-syscall, so this is the best fit.