Add API extensions
This patch adds API extensions to Trussed.  To avoid having to add a
type parameter to Request and Reply, we serialize the requests and
replies.  Extensions can be implemented by custom backends and are
dispatched by the Backends implementation provided to the Service.  An
additional type parameter is needed for ClientImplementation to express
the supported extensions.  See tests/ext.rs for an example
implementation.

This implementation has two limitations:
- As we cannot define a non-generic object-safe trait to execute
  extensions, Backends cannot just return a reference to the extension
  (as we do for the backends).  Instead it has to call the extension
  implementation directly.
- The serialization and deserialization of the requests and replies of
  course adds additional overhead.  An alternative would be to make
  Request and Reply generic over the set of extensions.